requirements.txt
pymongo
pandas
datetime
matplotlib.pyplot
math